Large refactor. Decoupling blocks from block template. WIP..
