+++ src/Events/Block_Templates/Block_Template_Contract.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TEC\Events\Editor\Full_Site;
+namespace TEC\Events\Block_Templates;
 
 use WP_Block_Template;
 
@@ -8,7 +8,6 @@ use WP_Block_Template;
  * Interface Block_Template_Contract
  */
 interface Block_Template_Contract {
-
 	/**
 	 * The Block ID.
 	 *
@@ -19,24 +18,6 @@ interface Block_Template_Contract {
 	public function id(): string;
 
 	/**
-	 * The Block slug.
-	 *
-	 * @since 6.2.7
-	 *
-	 * @return string
-	 */
-	public function slug(): string;
-
-	/**
-	 * Our namespace for a set of blocks.
-	 *
-	 * @since 6.2.7
-	 *
-	 * @return string
-	 */
-	public function get_namespace() : string;
-
-	/**
 	 * The getter for this template service to retrieve a hydrated WP_Block_Template.
 	 *
 	 * @since 6.2.7
@@ -44,13 +25,4 @@ interface Block_Template_Contract {
 	 * @return WP_Block_Template|null
 	 */
 	public function get_block_template(): ?WP_Block_Template;
-
-	/**
-	 * Handles rendering the template.
-	 *
-	 * @since 6.2.7
-	 *
-	 * @return string
-	 */
-	public function render(): string;
 }
