+++ src/Events/Block_Templates/Single_Event/Single_Block_Template.php
@@ -1,10 +1,11 @@
 <?php
 namespace TEC\Events\Block_Templates\Single_Event;
 
+use TEC\Events\Blocks\Single_Event\Block;
 use Tribe__Events__Main;
 use TEC\Common\Editor\Full_Site\Template_Utils;
 use WP_Block_Template;
-use TEC\Events\Editor\Full_Site\Block_Template_Contract;
+use TEC\Events\Block_Templates\Block_Template_Contract;
 
 /**
  * Class Single_Block_Template
@@ -13,23 +14,23 @@ use TEC\Events\Editor\Full_Site\Block_Template_Contract;
  *
  * @package TEC\Events\Block_Templates\Single_Event
  */
-class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements Block_Template_Contract {
+class Single_Block_Template implements Block_Template_Contract {
 	/**
-	 * @since 6.2.7
+	 * @since TBD
 	 *
-	 * @var string The namespace of this template.
+	 * @var Block The registered block for this template.
 	 */
-	protected $namespace = 'tec';
+	protected Block $block;
 
 	/**
-	 * Returns the name/slug of this block.
+	 * Constructor for Single Venue Block Template.
 	 *
-	 * @since 6.2.7
+	 * @since TBD
 	 *
-	 * @return string The name/slug of this block.
+	 * @param Block $block The registered Block for Single Venue.
 	 */
-	public function slug(): string {
-		return 'single-event';
+	public function __construct( Block $block ) {
+		$this->block = $block;
 	}
 
 	/**
@@ -40,36 +41,7 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 	 * @return string The WP Block Template ID.
 	 */
 	public function id(): string {
-		return $this->get_namespace() . '//' . $this->slug();
-	}
-
-	/**
-	 * Set the default attributes of this block.
-	 *
-	 * @since 6.2.7
-	 *
-	 * @return array<string,mixed> The array of default attributes.
-	 */
-	public function default_attributes(): array {
-		return [];
-	}
-
-	/**
-	 * Since we are dealing with a Dynamic type of Block we need a PHP method to render it.
-	 *
-	 * @since 6.2.7
-	 *
-	 * @param array $attributes The block attributes.
-	 *
-	 * @return string The block HTML.
-	 */
-	public function render( $attributes = [] ): string {
-		$args['attributes'] = $this->attributes( $attributes );
-
-		// Add the rendering attributes into global context.
-		tribe( 'events.editor.template' )->add_template_globals( $args );
-
-		return tribe( 'events.editor.template' )->template( [ 'blocks', $this->slug() ], $args, false );
+		return $this->block->get_namespace() . '//' . $this->block->slug();
 	}
 
 	/**
@@ -91,7 +63,7 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 			tribe_get_event_label_singular_lowercase()
 		);
 		$insert       = [
-			'post_name'    => $this->slug(),
+			'post_name'    => $this->block->slug(),
 			'post_title'   => $post_title,
 			'post_excerpt' => $post_excerpt,
 			'post_type'    => 'wp_template',
@@ -100,7 +72,7 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 				Tribe__Events__Main::instance()->plugin_path . '/src/Events/Block_Templates/Single_Event/templates/single-event.html'
 			) ),
 			'tax_input'    => [
-				'wp_theme' => $this->get_namespace()
+				'wp_theme' => $this->block->get_namespace()
 			]
 		];
 
@@ -116,7 +88,7 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 	 * @return null|WP_Block_Template The hydrated single events template object.
 	 */
 	public function get_block_template(): ?WP_Block_Template {
-		$wp_block_template = Template_Utils::find_block_template_by_post( $this->slug(), $this->get_namespace() );
+		$wp_block_template = Template_Utils::find_block_template_by_post( $this->block->slug(), $this->block->get_namespace() );
 
 		// If empty, this is our first time loading our Block Template. Let's create it.
 		if ( ! $wp_block_template ) {
@@ -128,8 +100,8 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 			do_action( 'tribe_log', 'error',
 				'Failed locating our WP_Block_Template for the Single Event Block', [
 					'method'    => __METHOD__,
-					'slug'      => $this->slug(),
-					'namespace' => $this->get_namespace()
+					'slug'      => $this->block->slug(),
+					'namespace' => $this->block->get_namespace()
 				] );
 		}
