+++ src/Events/Block_Templates/Single_Venue/Single_Block_Template.php
@@ -1,36 +1,38 @@
 <?php
 
-namespace TEC\Events\Editor\Full_Site\Venue;
+namespace TEC\Events\Block_Templates\Single_Venue;
 
+use TEC\Events\Blocks\Single_Venue\Block;
 use Tribe__Events__Main;
 use TEC\Common\Editor\Full_Site\Template_Utils;
 use WP_Block_Template;
-use TEC\Events\Editor\Full_Site\Block_Template_Contract;
+use TEC\Events\Block_Templates\Block_Template_Contract;
 
 /**
  * Class Single_Block_Template
  *
  * @since TBD
  *
- * @package TEC\Events\Editor\Full_Site\Venue
+ * @package TEC\Events\Block_Templates\Single_Venue
  */
-class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements Block_Template_Contract {
-	/**
-	 * @since TBD
-	 *
-	 * @var string The namespace of this template.
-	 */
-	protected $namespace = 'tec';
+class Single_Block_Template implements Block_Template_Contract {
 
 	/**
-	 * Returns the name/slug of this block.
+	 * @since TBD
+	 *
+	 * @var Block The registered block for this template.
+	 */
+	protected Block $block;
+
+	/**
+	 * Constructor for Single Venue Block Template.
 	 *
 	 * @since TBD
 	 *
-	 * @return string The name/slug of this block.
+	 * @param Block $block The registered Block for Single Venue.
 	 */
-	public function slug(): string {
-		return 'single-venue';
+	public function __construct( Block $block ) {
+		$this->block = $block;
 	}
 
 	/**
@@ -41,36 +43,7 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 	 * @return string The WP Block Template ID.
 	 */
 	public function id(): string {
-		return $this->get_namespace() . '//' . $this->slug();
-	}
-
-	/**
-	 * Set the default attributes of this block.
-	 *
-	 * @since TBD
-	 *
-	 * @return array<string,mixed> The array of default attributes.
-	 */
-	public function default_attributes(): array {
-		return [];
-	}
-
-	/**
-	 * Since we are dealing with a Dynamic type of Block we need a PHP method to render it.
-	 *
-	 * @since TBD
-	 *
-	 * @param array $attributes The block attributes.
-	 *
-	 * @return string The block HTML.
-	 */
-	public function render( $attributes = [] ): string {
-		$args['attributes'] = $this->attributes( $attributes );
-
-		// Add the rendering attributes into global context.
-		tribe( 'events.editor.template' )->add_template_globals( $args );
-
-		return tribe( 'events.editor.template' )->template( [ 'blocks', $this->slug() ], $args, false );
+		return $this->block->get_namespace() . '//' . $this->block->slug();
 	}
 
 	/**
@@ -92,16 +65,16 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 			tribe_get_venue_label_singular_lowercase()
 		);
 		$insert       = [
-			'post_name'    => $this->slug(),
+			'post_name'    => $this->block->slug(),
 			'post_title'   => $post_title,
 			'post_excerpt' => $post_excerpt,
 			'post_type'    => 'wp_template',
 			'post_status'  => 'publish',
 			'post_content' => Template_Utils::inject_theme_attribute_in_content( file_get_contents(
-				Tribe__Events__Main::instance()->plugin_path . '/src/Events/Blocks/Single_Venue_Template/templates/single-venue.html'
+				Tribe__Events__Main::instance()->plugin_path . '/src/Events/Block_Templates/Single_Venue/templates/single-venue.html'
 			) ),
 			'tax_input'    => [
-				'wp_theme' => $this->get_namespace()
+				'wp_theme' => $this->block->get_namespace()
 			]
 		];
 
@@ -117,7 +90,7 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 	 * @return null|WP_Block_Template The hydrated single events template object.
 	 */
 	public function get_block_template(): ?WP_Block_Template {
-		$wp_block_template = Template_Utils::find_block_template_by_post( $this->slug(), $this->get_namespace() );
+		$wp_block_template = Template_Utils::find_block_template_by_post( $this->block->slug(), $this->block->get_namespace() );
 
 		// If empty, this is our first time loading our Block Template. Let's create it.
 		if ( ! $wp_block_template ) {
@@ -129,8 +102,8 @@ class Single_Block_Template extends \Tribe__Editor__Blocks__Abstract implements 
 			do_action( 'tribe_log', 'error',
 				'Failed locating our WP_Block_Template for the Single Venue Block', [
 					'method'    => __METHOD__,
-					'slug'      => $this->slug(),
-					'namespace' => $this->get_namespace()
+					'slug'      => $this->block->slug(),
+					'namespace' => $this->block->get_namespace()
 				] );
 		}
