+++ src/Events/Blocks/Controller.php
@@ -5,8 +5,9 @@ namespace TEC\Events\Blocks;
 use TEC\Events\Block_Templates\Archive_Events\Archive_Block_Template;
 use TEC\Events\Block_Templates\Single_Event\Single_Block_Template;
 use TEC\Common\Contracts\Provider\Controller as Controller_Contract;
-use TEC\Events\Editor\Full_Site\Venue\Single_Block_Template as Single_Venue_Block_Template;
-// @todo Ditch this - does very little, seems odd considering lift of FS/Controller ?
+use TEC\Events\Block_Templates\Single_Venue\Single_Block_Template as Single_Venue_Block_Template;
+use TEC\Events\Blocks\Single_Venue\Block;
+
 /**
  * Class Controller
  *
@@ -40,7 +41,9 @@ class Controller extends Controller_Contract {
 	 * @since 6.2.7
 	 */
 	protected function add_actions() {
-
+		add_action( 'tribe_editor_register_blocks', [ $this, 'action_register_archive_template' ] );
+		add_action( 'tribe_editor_register_blocks', [ $this, 'action_register_single_event_template' ] );
+		add_action( 'tribe_editor_register_blocks', [ $this, 'register_single_venue_block' ] );
 	}
 
 	/**
@@ -49,13 +52,13 @@ class Controller extends Controller_Contract {
 	 * @since 6.2.7
 	 */
 	public function remove_actions() {
-
+		remove_action( 'tribe_editor_register_blocks', [ $this, 'action_register_archive_template' ] );
+		remove_action( 'tribe_editor_register_blocks', [ $this, 'action_register_single_event_template' ] );
+		remove_action( 'tribe_editor_register_blocks', [ $this, 'register_single_venue_block' ] );
 	}
 
-
-
 	/**
-	 * Registers the Events Archive template.
+	 * Registers the Events Archive block.
 	 *
 	 * @since 6.2.7
 	 */
@@ -64,11 +67,20 @@ class Controller extends Controller_Contract {
 	}
 
 	/**
-	 * Registers the Single Event template.
+	 * Registers the Single Event block.
 	 *
 	 * @since 6.2.7
 	 */
 	public function action_register_single_event_template() {
 		return $this->container->make( Single_Block_Template::class )->register();
+	}
+
+	/**
+	 * Registers the Single Venue block.
+	 *
+	 * @since TBD
+	 */
+	public function register_single_venue_block() {
+		return $this->container->make( Block::class )->register();
 	}
 }
