+++ src/Events/Blocks/Single_Venue/Block.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace TEC\Events\Blocks\Single_Venue;
+
+use Tribe__Events__Main;
+use TEC\Common\Editor\Full_Site\Template_Utils;
+use WP_Block_Template;
+use TEC\Events\Block_Templates\Block_Template_Contract;
+
+/**
+ * Class Block
+ *
+ * @since TBD
+ *
+ * @package TEC\Events\Blocks\Single_Venue
+ */
+class Block extends \Tribe__Editor__Blocks__Abstract {
+	/**
+	 * @since TBD
+	 *
+	 * @var string The namespace of this template.
+	 */
+	protected $namespace = 'tec';
+
+	/**
+	 * Returns the name/slug of this block.
+	 *
+	 * @since TBD
+	 *
+	 * @return string The name/slug of this block.
+	 */
+	public function slug(): string {
+		return 'single-venue';
+	}
+
+	/**
+	 * Set the default attributes of this block.
+	 *
+	 * @since TBD
+	 *
+	 * @return array<string,mixed> The array of default attributes.
+	 */
+	public function default_attributes(): array {
+		return [];
+	}
+
+	/**
+	 * Since we are dealing with a Dynamic type of Block we need a PHP method to render it.
+	 *
+	 * @since TBD
+	 *
+	 * @param array $attributes The block attributes.
+	 *
+	 * @return string The block HTML.
+	 */
+	public function render( $attributes = [] ): string {
+		$args['attributes'] = $this->attributes( $attributes );
+
+		// Add the rendering attributes into global context.
+		tribe( 'events.editor.template' )->add_template_globals( $args );
+
+		return tribe( 'events.editor.template' )->template( [ 'blocks', $this->slug() ], $args, false );
+	}
+}
