--- src/Events/Blocks/Single_Venue_Template/templates/single-venue.html
@@ -1,3 +1,0 @@
-<!-- wp:template-part {"slug":"header","tagName":"header"} /-->
-<!-- wp:tec/single-venue {} /-->
-<!-- wp:template-part {"slug":"footer","tagName":"footer"} /-->
